update file paths and move idxproptype file to #models folder
