+++ src/app/models/propTypes/idxPage/index.js
